[OpenMP][Offload][L0] Enable sycl wrapper for interop (#4976)
Changes from the original PR.
- Moved intel_win_dlfcn.h to correct location
- Added missing wrapper API check after loading wrapper
- Enabled Windows build
- Updated copyright year in omptarget.rc

+++ openmp/libomptarget/src/device.cpp
@@ -985,7 +985,7 @@ int32_t DeviceTy::isSupportedDevice(void *DeviceType) {
 #if INTEL_CUSTOMIZATION
 __tgt_interop *DeviceTy::createInterop(int32_t InteropContext,
                                        int32_t NumPrefers,
-                                       intptr_t *PreferIDs) {
+                                       int32_t *PreferIDs) {
   if (RTL->create_interop)
     return RTL->create_interop(RTLDeviceID, InteropContext, NumPrefers,
                                PreferIDs);
